requirements.txt
# Data Processing
pandas==2.3.0
numpy==2.2.0
openpyxl==3.1.5

# NLP Libraries
spacy
scikit-learn==1.6.1
nltk==3.9.1
sentence-transformers==3.2.1

# Web Scraping & APIs
requests==2.32.3
beautifulsoup4==4.13.0
python-docx==1.1.2
pypdf2==3.0.1  

# Database
# sqlite3 (built-in with Python)
sqlalchemy==2.0.36

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.2

# Testing
pytest==8.3.3
pytest-cov==6.0.0

# Utilities
tqdm==4.67.1
click==8.1.7